moved order creation fixture to conftest
