+++ demostore_automation/tests/backend/orders/test_change_order_status.py
@@ -12,7 +12,6 @@ Tests are parametrized using pytest.
 """
 import pytest
 import logging as logger
-from demostore_automation.tests.backend.orders.test_create_order_smoke import my_orders_smoke_setup
 
 
 pytestmark = [pytest.mark.orders, pytest.mark.order_status]
